4.0.0.8

+++ include/sciter-x-types.h
@@ -131,10 +131,10 @@ enum GFX_LAYER
     #define FALSE (0)
   #endif
 
-typedef unsigned int UINT;
+  typedef unsigned int UINT;
   typedef int INT;
   typedef unsigned long long UINT64;
-  typedef int INT64;
+  typedef long long INT64;
 
   typedef unsigned char BYTE;
   typedef char16_t WCHAR;
@@ -202,7 +202,7 @@ typedef unsigned int UINT;
   typedef unsigned int UINT;
   typedef int INT;
   typedef unsigned long long UINT64;
-  typedef int INT64;
+  typedef long long INT64;
 
   typedef unsigned char BYTE;
   typedef char16_t WCHAR;
